backend/pyproject.toml
[project]
name = "home-central-backend"
version = "0.1.0"
description = "Home Central API backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "supabase>=2.11.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
